# v: cleanup calls to Scope.find() methods