--- a/vlib/v/ast/scope.v
+++ b/vlib/v/ast/scope.v
@@ -49,13 +49,7 @@ pub fn (s &Scope) find(name string) ?ScopeObject {
 	if s == unsafe { nil } {
 		return none
 	}
-	if name in s.objects {
-		return unsafe { s.objects[name] }
-	}
-	if s.dont_lookup_parent() {
-		return none
-	}
-	for sc := unsafe { s.parent }; true; sc = sc.parent {
+	for sc := unsafe { s }; true; sc = sc.parent {
 		if name in sc.objects {
 			return unsafe { sc.objects[name] }
 		}