--- a/vlib/v/fmt/fmt.v
+++ b/vlib/v/fmt/fmt.v
@@ -2363,7 +2363,7 @@ pub fn (mut f Fmt) ident(node ast.Ident) {
 			}
 		}
 		if !is_local && !node.name.contains('.') && !f.inside_const {
-			if obj := f.file.global_scope.find_const('${f.cur_mod}.${node.name}') {
+			if _ := f.file.global_scope.find_const('${f.cur_mod}.${node.name}') {
 				const_name := node.name.all_after_last('.')
 				f.write(const_name)
 				if node.or_expr.kind == .block {