This PR aims optimize scope usage.

![image](https://github.com/user-attachments/assets/94333eb3-a6af-4c44-aa93-0bd5b53af2fd)

<sub><a href="https://huly.app/guest/vlang-66f40c4d-a476b54c67-771fdd?token=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJsaW5rSWQiOiI2NzMyMzMzZWU3Y2M1YTc4NTQyYWUxNDUiLCJndWVzdCI6InRydWUiLCJlbWFpbCI6IiNndWVzdEBoYy5lbmdpbmVlcmluZyIsIndvcmtzcGFjZSI6InctYWxleGFuZGVyLXZsYW5nLTY2ZjQwYzRkLWE0NzZiNTRjNjctNzcxZmRjIn0.y96tJAyBIV1FTedfa74LMUuzhSQVjuE-yZHW7T2DrT4">Huly&reg;: <b>V_0.6-21278</b></a></sub>