@spytheman 2024-11-12
After 9d6ecbd5512231ebc83f426174f14f8b86e9d9a2 , it is a bit better for both -prod and without it:
```
#0 10:45:16 ^ opt_scope ~/code/v>v run .github/workflows/compare_pr_to_master.v -prod
==================================================================================================
Current git branch: opt_scope, commit: 9d6ecbd
    Compiling new V executables from PR branch: opt_scope, commit: 9d6ecbd ...
CPU: 2.42s      Real: 2.62s     Elapsed: 0:02.62        RAM: 403480KB   ./v -o vnew1 cmd/v
CPU: 2.41s      Real: 2.62s     Elapsed: 0:02.62        RAM: 403516KB   ./vnew1 -o vnew2 cmd/v
CPU: 2.42s      Real: 2.60s     Elapsed: 0:02.60        RAM: 387868KB   ./vnew2 -no-parallel -o vnew cmd/v
CPU: 0.20s      Real: 0.22s     Elapsed: 0:00.22        RAM: 47464KB    ./vnew -no-parallel -o nhw_current.c examples/hello_world.v
CPU: 1.98s      Real: 2.18s     Elapsed: 0:02.18        RAM: 387884KB   ./vnew -no-parallel -o nv_current.c cmd/v
CPU: 98.22s     Real: 101.16s   Elapsed: 1:41.16        RAM: 430704KB   ./vnew -no-parallel -prod -o vnew_prod cmd/v
>Size of        nhw_current.c:     460215
>Size of         nv_current.c:    6841910
>Size of                 vnew:   10439720
>Size of            vnew_prod:    3885784
Switched to branch 'master'
Your branch is up to date with 'origin/master'.
==================================================================================================
    Compiling old V executables from branch: master, commit: 5092626 ...
CPU: 2.42s      Real: 2.62s     Elapsed: 0:02.62        RAM: 403408KB   ./v -o vold1 cmd/v
CPU: 2.40s      Real: 2.62s     Elapsed: 0:02.62        RAM: 403392KB   ./vold1 -o vold2 cmd/v
CPU: 2.37s      Real: 2.60s     Elapsed: 0:02.60        RAM: 387832KB   ./vold2 -no-parallel -o vold cmd/v
CPU: 0.20s      Real: 0.22s     Elapsed: 0:00.22        RAM: 47460KB    ./vnew -no-parallel -o ohw_master.c examples/hello_world.v
CPU: 1.97s      Real: 2.16s     Elapsed: 0:02.16        RAM: 387844KB   ./vold -no-parallel -o ov_master.c cmd/v
CPU: 98.34s     Real: 101.27s   Elapsed: 1:41.27        RAM: 430812KB   ./vold -no-parallel -prod -o vold_prod cmd/v
>Size of            vold_prod:    3889880
>Size of         ohw_master.c:     460215
>Size of          ov_master.c:    6842912
>Size of                 vold:   10440216
>Size of            vold_prod:    3889880
==================================================================================================
File sizes so far ...
>>>>>> size("  nhw_current.c") - size("   ohw_master.c") =     460215 -     460215 =          0
>>>>>> size("   nv_current.c") - size("    ov_master.c") =    6841910 -    6842912 =      -1002
>>>>>> size("           vnew") - size("           vold") =   10439720 -   10440216 =       -496
Switched to branch 'opt_scope'
==================================================================================================
    Measuring at PR branch: opt_scope, commit: 9d6ecbd ...
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1                                                                                  
    1   -0.5% 1.00x faster  40.9ms ± σ:   0.3ms,  40.5ms… 41.2ms `./vnew_prod -check-syntax           examples/hello_world.v`
 >  2         base          41.1ms ± σ:   0.2ms,  40.9ms… 41.3ms `./vold_prod -check-syntax           examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2                                                                                  
    1   -1.4% 1.01x faster  41.2ms ± σ:   0.1ms,  41.2ms… 41.3ms `./vnew_prod -check-syntax           examples/hello_world.v`
 >  2         base          41.8ms ± σ:   0.3ms,  41.5ms… 42.1ms `./vold_prod -check-syntax           examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3                                                                                  
    1   -1.4% 1.01x faster  41.2ms ± σ:   0.3ms,  40.9ms… 41.6ms `./vnew_prod -check-syntax           examples/hello_world.v`
 >  2         base          41.7ms ± σ:   0.2ms,  41.4ms… 41.9ms `./vold_prod -check-syntax           examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1                                                                                  
    1   -1.8% 1.02x faster  58.5ms ± σ:   0.4ms,  58.2ms… 59.1ms `./vnew_prod -check                  examples/hello_world.v`
 >  2         base          59.6ms ± σ:   0.4ms,  59.0ms… 59.9ms `./vold_prod -check                  examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2                                                                                  
    1   -1.6% 1.02x faster  59.1ms ± σ:   0.6ms,  58.2ms… 59.5ms `./vnew_prod -check                  examples/hello_world.v`
 >  2         base          60.0ms ± σ:   0.1ms,  59.9ms… 60.1ms `./vold_prod -check                  examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3                                                                                  
    1   -0.2% 1.00x faster  59.1ms ± σ:   0.2ms,  58.8ms… 59.3ms `./vnew_prod -check                  examples/hello_world.v`
 >  2         base          59.2ms ± σ:   0.6ms,  58.4ms… 59.7ms `./vold_prod -check                  examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1                                                                                  
    1   -0.6% 1.01x faster  85.6ms ± σ:   0.1ms,  85.5ms… 85.8ms `./vnew_prod -no-parallel -o nhw.c   examples/hello_world.v`
 >  2         base          86.2ms ± σ:   0.2ms,  86.0ms… 86.4ms `./vold_prod -no-parallel -o ohw.c   examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2                                                                                  
    1   -0.2% 1.00x faster  85.6ms ± σ:   0.3ms,  85.3ms… 85.9ms `./vnew_prod -no-parallel -o nhw.c   examples/hello_world.v`
 >  2         base          85.8ms ± σ:   0.5ms,  85.1ms… 86.2ms `./vold_prod -no-parallel -o ohw.c   examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3                                                                                  
    1   -0.3% 1.00x faster  85.6ms ± σ:   0.9ms,  84.4ms… 86.3ms `./vnew_prod -no-parallel -o nhw.c   examples/hello_world.v`
 >  2         base          85.9ms ± σ:   0.9ms,  84.7ms… 86.9ms `./vold_prod -no-parallel -o ohw.c   examples/hello_world.v`
>>>>>> size("          nhw.c") - size("          ohw.c") =     460215 -     460215 =          0
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1                                                                                  
    1   -0.7% 1.01x faster 114.3ms ± σ:   0.2ms, 114.1ms…114.5ms `./vnew_prod -no-parallel -o nhw.exe examples/hello_world.v`
 >  2         base         115.2ms ± σ:   0.1ms, 114.9ms…115.3ms `./vold_prod -no-parallel -o ohw.exe examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2                                                                                  
    1   -0.3% 1.00x faster 114.3ms ± σ:   0.3ms, 114.0ms…114.7ms `./vnew_prod -no-parallel -o nhw.exe examples/hello_world.v`
 >  2         base         114.6ms ± σ:   0.3ms, 114.3ms…115.0ms `./vold_prod -no-parallel -o ohw.exe examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3                                                                                  
 >  1         base         114.7ms ± σ:   0.4ms, 114.2ms…115.0ms `./vold_prod -no-parallel -o ohw.exe examples/hello_world.v`
    2   +0.1% 1.00x ~same~ 114.9ms ± σ:   0.1ms, 114.7ms…115.0ms `./vnew_prod -no-parallel -o nhw.exe examples/hello_world.v`
>>>>>> size("        nhw.exe") - size("        ohw.exe") =     726704 -     726704 =          0
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1                                                                                  
    1   -1.7% 1.02x faster 317.2ms ± σ:   1.0ms, 315.9ms…318.2ms `./vnew_prod -check-syntax           cmd/v`
 >  2         base         322.8ms ± σ:   1.6ms, 320.5ms…324.0ms `./vold_prod -check-syntax           cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2                                                                                  
    1   -1.1% 1.01x faster 317.6ms ± σ:   1.4ms, 315.6ms…318.7ms `./vnew_prod -check-syntax           cmd/v`
 >  2         base         321.1ms ± σ:   0.7ms, 320.2ms…321.7ms `./vold_prod -check-syntax           cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3                                                                                  
    1   -1.0% 1.01x faster 318.2ms ± σ:   1.0ms, 316.9ms…319.4ms `./vnew_prod -check-syntax           cmd/v`
 >  2         base         321.4ms ± σ:   0.4ms, 320.9ms…321.9ms `./vold_prod -check-syntax           cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1                                                                                  
    1   -0.6% 1.01x faster 530.8ms ± σ:   0.8ms, 529.6ms…531.5ms `./vnew_prod -check                  cmd/v`
 >  2         base         534.2ms ± σ:   1.1ms, 532.7ms…535.4ms `./vold_prod -check                  cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2                                                                                  
    1   -0.7% 1.01x faster 530.0ms ± σ:   0.5ms, 529.2ms…530.4ms `./vnew_prod -check                  cmd/v`
 >  2         base         533.8ms ± σ:   1.4ms, 531.9ms…534.8ms `./vold_prod -check                  cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3                                                                                  
    1   -1.0% 1.01x faster 530.1ms ± σ:   0.2ms, 529.8ms…530.3ms `./vnew_prod -check                  cmd/v`
 >  2         base         535.4ms ± σ:   0.6ms, 534.8ms…536.2ms `./vold_prod -check                  cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1                                                                                  
    1   -0.6% 1.01x faster 840.7ms ± σ:   0.6ms, 840.0ms…841.3ms `./vnew_prod -no-parallel -o nv.c    cmd/v`
 >  2         base         846.0ms ± σ:   1.0ms, 845.2ms…847.4ms `./vold_prod -no-parallel -o ov.c    cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2                                                                                  
    1   -0.7% 1.01x faster 840.9ms ± σ:   0.3ms, 840.4ms…841.1ms `./vnew_prod -no-parallel -o nv.c    cmd/v`
 >  2         base         847.1ms ± σ:   0.9ms, 846.4ms…848.3ms `./vold_prod -no-parallel -o ov.c    cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3                                                                                  
    1   -0.6% 1.01x faster 840.5ms ± σ:   0.4ms, 840.0ms…840.9ms `./vnew_prod -no-parallel -o nv.c    cmd/v`
 >  2         base         845.3ms ± σ:   1.2ms, 843.9ms…846.8ms `./vold_prod -no-parallel -o ov.c    cmd/v`
>>>>>> size("           nv.c") - size("           ov.c") =    6841910 -    6841910 =          0
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1                                                                                  
    1   -0.2% 1.00x faster 1275.9ms ± σ:   1.9ms, 1273.6ms…1278.2ms `./vnew_prod -no-parallel -o nv.exe  cmd/v`
 >  2         base         1278.1ms ± σ:   0.8ms, 1277.5ms…1279.2ms `./vold_prod -no-parallel -o ov.exe  cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2                                                                                  
    1   -0.4% 1.00x faster 1272.7ms ± σ:   1.3ms, 1270.9ms…1274.1ms `./vnew_prod -no-parallel -o nv.exe  cmd/v`
 >  2         base         1277.3ms ± σ:   0.7ms, 1276.5ms…1278.3ms `./vold_prod -no-parallel -o ov.exe  cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3                                                                                  
    1   -0.5% 1.01x faster 1273.2ms ± σ:   0.9ms, 1271.9ms…1273.9ms `./vnew_prod -no-parallel -o nv.exe  cmd/v`
 >  2         base         1279.6ms ± σ:   1.3ms, 1278.2ms…1281.3ms `./vold_prod -no-parallel -o ov.exe  cmd/v`
>>>>>> size("         nv.exe") - size("         ov.exe") =   10439736 -   10439736 =          0
Done. Total time: 460.171234865 s.
==================================================================================================
Final summary for file diff sizes on their own branches:
>>>>>> size("  nhw_current.c") - size("   ohw_master.c") =     460215 -     460215 =          0
>>>>>> size("   nv_current.c") - size("    ov_master.c") =    6841910 -    6842912 =      -1002
>>>>>> size("           vnew") - size("           vold") =   10439720 -   10440216 =       -496
>>>>>> size("      vnew_prod") - size("      vold_prod") =    3885784 -    3889880 =      -4096
==================================================================================================
Final summary for file diff sizes for generated files on the *current* branch:
>>>>>> size("          nhw.c") - size("          ohw.c") =     460215 -     460215 =          0
>>>>>> size("           nv.c") - size("           ov.c") =    6841910 -    6841910 =          0
>>>>>> size("        nhw.exe") - size("        ohw.exe") =     726704 -     726704 =          0
>>>>>> size("         nv.exe") - size("         ov.exe") =   10439736 -   10439736 =          0
#0 10:53:04 ^ opt_scope ~/code/v>
#0 10:56:31 ^ opt_scope ~/code/v>
#0 10:56:31 ^ opt_scope ~/code/v>
#0 10:56:31 ^ opt_scope ~/code/v>
#0 10:56:31 ^ opt_scope ~/code/v>v run .github/workflows/compare_pr_to_master.v 
==================================================================================================
Current git branch: opt_scope, commit: 9d6ecbd
    Compiling new V executables from PR branch: opt_scope, commit: 9d6ecbd ...
CPU: 2.39s      Real: 2.62s     Elapsed: 0:02.62        RAM: 403480KB   ./v -o vnew1 cmd/v
CPU: 2.42s      Real: 2.62s     Elapsed: 0:02.62        RAM: 403516KB   ./vnew1 -o vnew2 cmd/v
CPU: 2.38s      Real: 2.60s     Elapsed: 0:02.60        RAM: 387868KB   ./vnew2 -no-parallel -o vnew cmd/v
CPU: 0.20s      Real: 0.22s     Elapsed: 0:00.22        RAM: 47464KB    ./vnew -no-parallel -o nhw_current.c examples/hello_world.v
CPU: 1.97s      Real: 2.17s     Elapsed: 0:02.17        RAM: 387884KB   ./vnew -no-parallel -o nv_current.c cmd/v
>Size of        nhw_current.c:     460215
>Size of         nv_current.c:    6841910
>Size of                 vnew:   10439720
Switched to branch 'master'
Your branch is up to date with 'origin/master'.
==================================================================================================
    Compiling old V executables from branch: master, commit: 5092626 ...
CPU: 2.41s      Real: 2.62s     Elapsed: 0:02.62        RAM: 403408KB   ./v -o vold1 cmd/v
CPU: 2.42s      Real: 2.62s     Elapsed: 0:02.62        RAM: 403392KB   ./vold1 -o vold2 cmd/v
CPU: 2.38s      Real: 2.60s     Elapsed: 0:02.60        RAM: 387832KB   ./vold2 -no-parallel -o vold cmd/v
CPU: 0.19s      Real: 0.22s     Elapsed: 0:00.22        RAM: 47460KB    ./vnew -no-parallel -o ohw_master.c examples/hello_world.v
CPU: 2.00s      Real: 2.17s     Elapsed: 0:02.17        RAM: 387844KB   ./vold -no-parallel -o ov_master.c cmd/v
>Size of         ohw_master.c:     460215
>Size of          ov_master.c:    6842912
>Size of                 vold:   10440216
==================================================================================================
File sizes so far ...
>>>>>> size("  nhw_current.c") - size("   ohw_master.c") =     460215 -     460215 =          0
>>>>>> size("   nv_current.c") - size("    ov_master.c") =    6841910 -    6842912 =      -1002
>>>>>> size("           vnew") - size("           vold") =   10439720 -   10440216 =       -496
Switched to branch 'opt_scope'
==================================================================================================
    Measuring at PR branch: opt_scope, commit: 9d6ecbd ...
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1                                                                                  
 >  1         base         111.8ms ± σ:   0.3ms, 111.3ms…112.1ms `./vold -check-syntax           examples/hello_world.v`
    2   +0.0% 1.00x ~same~ 111.8ms ± σ:   0.6ms, 111.0ms…112.3ms `./vnew -check-syntax           examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2                                                                                  
    1   -1.0% 1.01x faster 111.8ms ± σ:   0.7ms, 110.8ms…112.4ms `./vnew -check-syntax           examples/hello_world.v`
 >  2         base         112.9ms ± σ:   0.2ms, 112.8ms…113.2ms `./vold -check-syntax           examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3                                                                                  
    1   -0.1% 1.00x ~same~ 112.2ms ± σ:   0.1ms, 112.0ms…112.3ms `./vnew -check-syntax           examples/hello_world.v`
 >  2         base         112.3ms ± σ:   0.6ms, 111.5ms…113.0ms `./vold -check-syntax           examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1                                                                                  
    1   -1.0% 1.01x faster 164.2ms ± σ:   0.1ms, 164.1ms…164.3ms `./vnew -check                  examples/hello_world.v`
 >  2         base         165.9ms ± σ:   0.5ms, 165.2ms…166.2ms `./vold -check                  examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2                                                                                  
    1   -0.6% 1.01x faster 164.6ms ± σ:   0.2ms, 164.3ms…164.9ms `./vnew -check                  examples/hello_world.v`
 >  2         base         165.6ms ± σ:   0.3ms, 165.3ms…165.8ms `./vold -check                  examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3                                                                                  
    1   -0.2% 1.00x faster 164.8ms ± σ:   0.1ms, 164.7ms…165.0ms `./vnew -check                  examples/hello_world.v`
…